splitting out searchmedia from app
